+++ CHANGELOG.md
@@ -32,7 +32,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.2.1] - 2020-11-03
 
+### Added
+- Condition for comparing products based on the category tree of the first product selected for comparison, this condition renders the checkbox unclickable for products that do not match the category of the first selected product.
+
 ## [0.2.0] - 2020-10-08
+
 
 ## [0.1.0] - 2020-09-29